requirements.txt
# Main requirements
sanic>=22.12.0
asyncpg>=0.27.0
bcrypt>=4.0.1
PyJWT>=2.8.0
python-dotenv>=1.0.0

# Testing 
pytest>=7.2.0
pytest-asyncio>=0.21.0
httpx>=0.24.0